Multi-Agent-AI-Research-Assistant-Summarizer-With-MCP-main/requirements.txt
streamlit
fastmcp
langchain
langchain-core
langchain-google-genai
langchain-mcp-adapters
langgraph
python-dotenv
pydantic
mcp-use
